mainページの要素をファイル分けしました．
