+++ features/main/components/index.tsx
@@ -1,0 +1,5 @@
+import GitCalendar from "./GitCalendar/GitCalendar";
+import MarqueeWidget from "./Marquee/Marquee";
+import Notice from "./Notice/Notice";
+
+export { GitCalendar, MarqueeWidget, Notice };
